Refactor - add features
