+++ src/features/categories/private/categories.types.ts
